static method & fix tests
