+++ tests/jolocomRegistry/commit.test.ts
@@ -4,7 +4,8 @@ import * as sinonChai from 'sinon-chai'
 import { createJolocomRegistry } from '../../ts/registries/jolocomRegistry'
 import { IdentityWallet } from '../../ts/identityWallet/identityWallet'
 import { Identity } from '../../ts/identity/identity'
-import { SoftwareKeyProvider, IKeyDerivationArgs } from '../../ts/vaultedKeyProvider/softwareProvider'
+import { SoftwareKeyProvider } from '../../ts/vaultedKeyProvider/softwareProvider'
+import { IKeyDerivationArgs } from '../../ts/vaultedKeyProvider/types'
 import { testPrivateEthereumKey, testSeed } from '../data/keys.data'
 import { DidDocument } from '../../ts/identity/didDocument/didDocument'
 import { didDocumentJSON, mockDid, mockIpfsHash } from '../data/didDocument.data'
