+++ ts/identityWallet/types.ts
@@ -1,5 +1,5 @@
 import { Identity } from '../identity/identity'
-import { IVaultedKeyProvider } from '../vaultedKeyProvider/softwareProvider'
+import { IVaultedKeyProvider } from '../vaultedKeyProvider/types'
 import { IKeyMetadata } from '../credentials/signedCredential/types'
 
 export interface IIdentityWalletCreateArgs {
